Remove now-unused logic about "volatile" context properties.
In light of #206, many more expressions are "volatile" in the sense
explained in the removed comments: they must be evaluated before yielding,
because they might evaluate to a different value after the yield. In fact,
as far as we can reasonbly prove, any expression that's not a Literal is
potentially volatile.