Don't invoke .return when one iteration of an async generator fails.
This seemed like the right thing to do, but the included tests show that
it can lead to unexpected (and probably undesired) behavior. It's up to
the consumer of the AsyncIterator to decide how to handle rejections, and
the consumer can easily call .throw or .return if that makes sense.