Don't coerce falsy this values to null in regeneratorRuntime.wrap.
In non-strict mode, values like false and undefined get coerced to objects
when passed as the first argument to .call or .apply, which is almost
never what you want. In strict mode, the specified value is left
untouched, which is probably a saner behavior. With this change,
Regenerator will no longer express any opinion about the matter.

Fixes #217.