Make AsyncIterator compatible with leebyron/async-to-gen.
The `runtime.awrap` method now returns a special `{ __await: value }`
structure instead of `new AwaitArgument(value)` to indicate that a value
yielded from an async generator should be awaited.

Checking `hasOwn.call(value, "__await")` is functionally equivalent to
checking `value instanceof AwaitArgument`, but it's safer because it
allows other code generators to compile async generators without having
access to the `AwaitArgument` class. Note in particular that "other code
generators" might mean other copies of the regenerator-runtime library,
each with their own version of the `AwaitArgument` class.

The `AwaitArgument` wrapper class was never more than an internal
implementation detail of `runtime.awrap`, so this is not a breaking change
for any existing client code of regenerator-runtime.