Stop using RegExp to detect possible generator/async syntax.
The regenerator package has historically benefitted from checking for
`function*` and/or `async` in the source text before compilation, since
nothing needed to be compiled if that syntax was not present.

However, after PR #371, which added support for class generator methods,
this heuristic does not work as well, since the `function` keyword no
longer has to precede the `*` character. Instead, I think it makes more
sense to compile code unconditionally.

Note that this behavior only applies to the regenerator package, which is
not used by Babel.