Add LICENSE files to distributed packages
The published packages were missing the complete LICENSE file,
since "the root directory of this source tree" was not present
in the distributed version.

This commit includes the files inside the individual packages
in order to appease corporate legal departments with extremely
strict FOSS requirements and fail to see SPDX declaration in
the package.json as an "authoritative enough" indication of
author's intent.